+++ src/main/java/org/monarchinitiative/hpoannotqc/cmd/BigFileCommand.java
@@ -1,6 +1,8 @@
 package org.monarchinitiative.hpoannotqc.cmd;
 
 
+import com.beust.jcommander.Parameter;
+import com.beust.jcommander.Parameters;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.monarchinitiative.hpoannotqc.io.PhenotypeDotHpoaFileWriter;
@@ -20,32 +22,31 @@ import java.util.List;
  * with the {@link DownloadCommand}).
  * @author <a href="mailto:peter.robinson@jax.org">Peter Robinson</a>
  */
+@Parameters(commandDescription = "Create phenotype.hpoa file")
 public class BigFileCommand implements Command {
     private static final Logger logger = LogManager.getLogger();
-    /** Path to directory with the ca. 7000 HPO Annotation files ("small files"). */
-    private final String hpoAnnotationFileDirectory;
     /** Path to the {@code hp.obo} file. */
     private final String hpOboPath;
     /** Path to the downloaded Orphanet XML file */
     private final String orphanetXMLpath;
-    /** Should usually be phenotype.hpoa, may also include path */
-    private final String outputFilePath;
-    /** path to the omit-list.txt file, which is located with the small files in the same directory */
-    private final String omitPath;
+    /** Directory with hp.obo and en_product>HPO.xml files. */
+    @Parameter(names={"-d","--data"}, description ="directory to download data (default: data)" )
+    private String downloadDirectory="data";
 
-    /**
-     * Command to create the V2 bigfile from the various small files
-     * @param hpopath path to hp.obo
-     * @param dir directory with the small files
-     * @param orphaXML path to the Orphanet XML file
-     * @param outpath path to the new output file (phenotype.hpoa)
-     */
-    public BigFileCommand(String hpopath, String dir, String orphaXML, String outpath) {
-        hpOboPath=hpopath;
-        hpoAnnotationFileDirectory =dir;
-        orphanetXMLpath=orphaXML;
-        outputFilePath=outpath;
-        omitPath=String.format("%s%s%s", hpoAnnotationFileDirectory,File.separator,"omit-list.txt");
+    @Parameter(names={"-a","--annot"},description = "Path to directory with the ca. 7000 HPO Annotation files ", required = true)
+    private String hpoAnnotationFileDirectory;
+    /** Should usually be phenotype.hpoa, may also include path */
+    @Parameter(names={"-o","--output"},description="name of output file")
+    private String outputFilePath="phenotype.hpoa";
+
+    @Parameter(names="--tolerant",description = "tolerant mode (update obsolte term ids if possible)")
+    private boolean tolerant=true;
+
+    /** Command to create the{@code phenotype.hpoa} file from the various small HPO Annotation files. */
+    public BigFileCommand() {
+        hpOboPath=String.format("%s%s%s",downloadDirectory,File.separator, "hp.obo" );
+        orphanetXMLpath=String.format("%s%s%s",downloadDirectory,File.separator, "en_product4_HPO.xml" );
+
     }
 
     @Override
@@ -60,12 +61,15 @@ public class BigFileCommand implements Command {
             logger.fatal("Unable to recover, stopping execution");
             return;
         }
+        // path to the omit-list.txt file, which is located with the small files in the same directory
+        String omitPath=String.format("%s%s%s", hpoAnnotationFileDirectory,File.separator,"omit-list.txt");
+        System.err.println("[INFO] annotation="+hpoAnnotationFileDirectory);
         try {
             // 1. Get the HPO project annotation files
             HpoAnnotationFileIngestor annotationFileIngestor = new HpoAnnotationFileIngestor(hpoAnnotationFileDirectory, omitPath, ontology);
             List<HpoAnnotationModel> hpoFileEntryList = annotationFileIngestor.getV2SmallFileEntries();
             // 2. Get the Orphanet annotation file
-            OrphanetXML2HpoDiseaseModelParser orphaParser = new OrphanetXML2HpoDiseaseModelParser(this.orphanetXMLpath, ontology);
+            OrphanetXML2HpoDiseaseModelParser orphaParser = new OrphanetXML2HpoDiseaseModelParser(this.orphanetXMLpath, ontology, tolerant);
             List<HpoAnnotationModel> orphaFileEntryList = orphaParser.getOrphanetDiseaseModels();
             // 3. Combine both and output to phenotype.hpoa ("big file")
             PhenotypeDotHpoaFileWriter writer = new PhenotypeDotHpoaFileWriter(ontology, hpoFileEntryList, orphaFileEntryList, outputFilePath);