+++ src/main/java/org/monarchinitiative/hpoannotqc/exception/HpoAnnotationModelException.java
@@ -4,7 +4,7 @@ import org.monarchinitiative.phenol.base.PhenolException;
 
 /**
  * Exceptions of this class are thrown if there are quality control issues with the
- * generation of a HpoAnnotationModel or HpoAnnotationFileEntry (single line of an HPO annotation file)
+ * generation of a HpoAnnotationModel or HpoAnnotationEntry (single line of an HPO annotation file)
  * @author Peter Robinson
  */
 public class HpoAnnotationModelException extends PhenolException {